Lesson 5
Alter components to be registered locally. Alter plan and plan-picker component objects to be encapsulated in PlanComponent and PlanPickerComponent variables. Add components property to PlanPickerComponent object and pass plan as a property and the PlanPicker object as the value. Add components property to the Vue instance. Pass in plan as the property and the PlanPickerComponent variable as the value.