Lesson 6
Add a selectedPlan property to the PlanPickerComponent object. Add methods attribute to PlanPickerComponent object. Add selectPlan method that sets the selectedPlan property in data to the plan. Add methods property to PlanComponent object. Create a select method that creates a custom event using $emit to pass the name of the plan selected to the PlanPicker component. Add computed property to PlanComponent. Add isSelected method to return a boolean determining if the current plan is the selected plan. Add a listener for the custom select event to the plan in the plan-picker-template. Use the v-on directive to listen for a click event on the plan template to call the select method. Use the v-bind directive to set the selectedPlan attribute to the boolean value returned from isSelected. Use v-bind to add a class of active-plan if isSelected returns true for the current plan. Remove background properties in css that call missing images.